fix(MissionList): Moving the Cards into their own component
